cleaning up, adding unit test, refactoring code
