+++ deepchem/models/torch_models/__init__.py
@@ -10,7 +10,7 @@ from deepchem.models.torch_models.pagtn import Pagtn, PagtnModel
 from deepchem.models.torch_models.mat import MAT, MATModel
 from deepchem.models.torch_models.megnet import MEGNetModel
 from deepchem.models.torch_models.normalizing_flows_pytorch import NormalizingFlow
-from deepchem.models.torch_models.layers import LinearModule, CNNModule, CombineMeanStd, WeightedLinearCombo, AtomicConvolution, NeighborList
+from deepchem.models.torch_models.layers import CNNModule, CombineMeanStd, WeightedLinearCombo, AtomicConvolution, NeighborList
 from deepchem.models.torch_models.cnn import CNN
 from deepchem.models.torch_models.dmpnn import DMPNN, DMPNNModel
 from deepchem.models.torch_models.neuralode import NeuralODE, NeuralODEModel
